Change amount and price columns types from string to double and add total price as a field in expense table to show the total price for each price in the main expenses screen

+++ app/src/main/java/com/sobky/expensestracking/ui/expenseitemdetails/ExpenseItemInfoFragment.kt
@@ -149,7 +149,7 @@ class ExpenseItemInfoFragment : Fragment() {
 
         binding.etExpenseItemInfoAmount.doOnTextChanged { inputText, _, _, _ ->
             if (inputText.toString() != viewModel.expenseItem?.value?.amount.toString().trim()) {
-                viewModel.expenseItem?.value?.amount = inputText.toString()
+                viewModel.expenseItem?.value?.amount = inputText.toString().toDouble()
                 viewModel.updateExpenseItem()
             }
         }
@@ -165,7 +165,7 @@ class ExpenseItemInfoFragment : Fragment() {
 
         binding.etExpenseItemInfoPrice.doOnTextChanged { inputText, _, _, _ ->
             if (inputText.toString() != viewModel.expenseItem?.value?.price.toString().trim()) {
-                viewModel.expenseItem?.value?.price = inputText.toString()
+                viewModel.expenseItem?.value?.price = inputText.toString().toDouble()
                 viewModel.updateExpenseItem()
             }
         }
